+++ providers/core.storage.js
